Rollbacks, Codestyle, Exception and Close FactoryаÑ fixed

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoHibernateImpl.java
@@ -1,11 +1,12 @@
 package jm.task.core.jdbc.dao;
 
+import com.mysql.cj.xdevapi.SessionFactory;
 import jm.task.core.jdbc.model.User;
 import jm.task.core.jdbc.util.Util;
+import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class UserDaoHibernateImpl implements UserDao {
@@ -15,15 +16,17 @@ public class UserDaoHibernateImpl implements UserDao {
 
     @Override
     public void createUsersTable() {
-            try (Session session = Util.getSessionFactory().getCurrentSession()) {
-                session.beginTransaction();
-                session.createSQLQuery("CREATE TABLE IF NOT EXISTS users (" +
-                        "id BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY," +
-                        " name VARCHAR(50) NOT NULL," +
-                        " lastName VARCHAR(50) NOT NULL," +
-                        " age TINYINT NOT NULL)").executeUpdate();
-                session.getTransaction().commit();
-            }
+        try (Session session = Util.getSessionFactory().getCurrentSession()) {
+            session.beginTransaction();
+            session.createSQLQuery("CREATE TABLE IF NOT EXISTS users (" +
+                    "id BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY," +
+                    " name VARCHAR(50) NOT NULL," +
+                    " lastName VARCHAR(50) NOT NULL," +
+                    " age TINYINT NOT NULL)").executeUpdate();
+            session.getTransaction().commit();
+        } catch (HibernateException e) {
+            System.out.println(e.getMessage());
+        }
     }
 
     @Override
@@ -32,45 +35,71 @@ public class UserDaoHibernateImpl implements UserDao {
             session.beginTransaction();
             session.createSQLQuery("DROP TABLE IF EXISTS users").executeUpdate();
             session.getTransaction().commit();
+        } catch (HibernateException e) {
+            System.out.println(e.getMessage());
         }
     }
 
     @Override
     public void saveUser(String name, String lastName, byte age) {
-        try (Session session = Util.getSessionFactory().getCurrentSession()) {
+        Session session = Util.getSessionFactory().getCurrentSession();
+        try {
             User user = new User(name, lastName, age);
             session.beginTransaction();
             session.saveOrUpdate(user);
             session.getTransaction().commit();
+        } catch (HibernateException e) {
+            session.getTransaction().rollback();
+            System.out.println(e.getMessage());
+        } finally {
+            session.close();
         }
     }
 
     @Override
     public void removeUserById(long id) {
-        try (Session session = Util.getSessionFactory().getCurrentSession()) {
+        Session session = Util.getSessionFactory().getCurrentSession();
+        try {
             session.beginTransaction();
-            User userDelete = session.get(User.class, id);
-            session.delete(userDelete);
+            if (session.get(User.class, id) != null) {
+                User userDelete = session.get(User.class, id);
+                session.delete(userDelete);
+            }
             session.getTransaction().commit();
+        } catch (HibernateException e) {
+            session.getTransaction().rollback();
+            System.out.println(e.getMessage());
+        } finally {
+            session.close();
         }
     }
 
     @Override
     public List<User> getAllUsers() {
+        List<User> users = null;
         try (Session session = Util.getSessionFactory().getCurrentSession()) {
             session.beginTransaction();
-            List<User> users = session.createQuery("from User").list();
+            users = session.createQuery("from User").list();
             session.getTransaction().commit();
             return users;
+        } catch (HibernateException e) {
+            System.out.println(e.getMessage());
         }
+        return users;
     }
 
     @Override
     public void cleanUsersTable() {
-        try (Session session = Util.getSessionFactory().getCurrentSession()) {
+        Session session = Util.getSessionFactory().getCurrentSession();
+        try {
             session.beginTransaction();
             session.createQuery("delete from User").executeUpdate();
             session.getTransaction().commit();
+        } catch (HibernateException e) {
+            session.getTransaction().rollback();
+            System.out.println(e.getMessage());
+        } finally {
+            session.close();
         }
     }
 }
